backend/requirements.txt
fastapi
uvicorn
sqlalchemy
openpyxl
pywebview
requests
python-multipart
PyQt5
bs4
selenium
PyJWT==2.8.0
python-dotenv==1.0.0
bcrypt==4.1.2
pytest
httpx
pytest-asyncio
pre-commit
black
flake8
isort
passlib
